QALI-515 Statistics factory
Refactors the structure for the snapshots, fixes tests
