+++ gradle.properties
@@ -18,7 +18,7 @@ kapt.include.compile.classpath=false
 # https://youtrack.jetbrains.com/issue/KT-33889
 kapt.incremental.apt=false
 org.gradle.jvmargs=-Xmx2g -XX:MaxMetaspaceSize=512m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
-pluginPlatformVersion=0.10.1-SNAPSHOT
+pluginPlatformVersion=0.12.0-SNAPSHOT
 # In order to reduce the number of cores used to run the tests, you can set a low value here (like 2).
 # Leave it empty to use the available physical cores.
 testNumCpuCore=
