+++ graphite/build.gradle.kts
@@ -62,6 +62,7 @@ val ktorVersion = "2.2.4"
 
 dependencies {
     implementation(platform("io.qalipsis:qalipsis-plugin-platform:${pluginPlatformVersion}"))
+    implementation("io.ktor:ktor-client-java:2.2.4")
     compileOnly("io.aeris-consulting:catadioptre-annotations")
     compileOnly("io.micronaut:micronaut-runtime")
     compileOnly("org.jetbrains.kotlinx:kotlinx-coroutines-core")
