+++ project.version
@@ -1,1 +1,1 @@
-0.10.1-SNAPSHOT
+0.12.0-SNAPSHOT
